small cleanup and fix
fixed a bug- temporary buffers in render passes were being cleared to black before being rendered to, an unnecessary op that slowed things down :: added a way to reset the base time of ISFScene, allowing control over the rendering/antimation time :: added some basic utilities to simplify some cross-platform ops

+++ VVGL/include/VVGL_Time.hpp
@@ -87,7 +87,7 @@ struct Timestamp	{
 	
 	
 	friend inline ostream & operator<<(ostream & os, const Timestamp & rs)	{
-		os << "<Timestamp " << rs.getTimeInSeconds() << ">";
+		os << "<Timestamp " << double(rs.getTimeInSeconds()) << ">";
 		return os;
 	}
 	inline Timestamp operator-(const Timestamp & n) const	{
